feat(dashboard): ui change and main added
